# Project_1/micro-loan-pwa/openapi.yaml
openapi: 3.0.3
info:
  title: Micro-Loan Calculator API
  description: |
    Backend API for the Micro-Loan Calculator PWA.
    
    Supports:
    - Loan calculations with validation
    - Cloud sync for loan history and savings goals
    - Financial literacy content delivery
    - User authentication (optional)
    - Lender database (future)
    
    PRD Reference: Sections 3.1-3.4, Phase 3 requirements
  version: 1.0.0
  contact:
    name: API Support
    email: support@microloan.app
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.microloan.app/api
    description: Production server
...
paths:
...
  /loans/calculate:
    post:
      tags:
        - Loans
      summary: Calculate loan details
      description: |
        Performs loan calculation with validation.
        Supports flat, reducing balance, and compound interest types.
        Returns predatory loan warning if APR > 50%.
        PRD Reference: Section 3.1
      operationId: calculateLoan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanCalculationRequest'
            examples:
              reducing_balance:
                summary: Reducing balance loan
                value:
                  amount: 1000
                  interestRate: 12
                  interestType: reducing
                  termLength: 12
                  termUnit: months
                  fees: 50
                  currency: USD
              flat_rate:
                summary: Flat rate loan
                value:
                  amount: 5000
                  interestRate: 15
                  interestType: flat
                  termLength: 2
                  termUnit: years
                  fees: 0
                  currency: KES
      responses:
        '200':
          description: Calculation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanCalculationResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /loans:
    get:
      tags:
        - Loans
      summary: Get user's loan history
      description: Retrieve all saved loan calculations for the authenticated user
      operationId: getLoans
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of loan calculations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanHistoryResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Loans
      summary: Save a loan calculation
      description: Save a loan calculation to the user's history (cloud sync)
      operationId: saveLoan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveLoanRequest'
      responses:
        '201':
          description: Loan saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanRecord'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /loans/{id}:
    get:
      tags:
        - Loans
      summary: Get a specific loan calculation
      operationId: getLoanById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Loan details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanRecord'
        '404':
          description: Loan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Loans
      summary: Delete a loan calculation
      operationId: deleteLoan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Loan deleted successfully
        '404':
          description: Loan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================
  # SAVINGS ENDPOINTS
  # ==================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoanCalculationRequest:
      type: object
      required:
        - amount
        - interestRate
        - interestType
        - termLength
        - termUnit
      properties:
        amount:
          type: number
          minimum: 1
          description: Loan principal amount
          example: 1000
        interestRate:
          type: number
          minimum: 0
          maximum: 500
          description: Annual interest rate as percentage
          example: 12
        interestType:
          type: string
          enum: [flat, reducing, compound]
          description: Type of interest calculation
          example: reducing
        termLength:
          type: integer
          minimum: 1
          description: Loan term length
          example: 12
        termUnit:
          type: string
          enum: [weeks, months, years]
          description: Unit for term length
          example: months
        fees:
          type: number
          minimum: 0
          default: 0
          description: Additional fees
          example: 50
        currency:
          type: string
          default: USD
          enum: [USD, KES, NGN, INR, PHP]
          description: Currency code
          example: USD
    LoanCalculationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            monthlyPayment:
              type: number
              description: Monthly payment amount
              example: 88.85
            totalInterest:
              type: number
              description: Total interest over loan term
              example: 66.19
            totalCost:
              type: number
              description: Total amount to repay (principal + interest + fees)
              example: 1116.19
            effectiveAPR:
              type: number
              description: Effective annual percentage rate
              example: 13.24
            interestType:
              type: string
              enum: [flat, reducing, compound]
            input:
              $ref: '#/components/schemas/LoanCalculationRequest'
        warning:
          type: object
          nullable: true
          properties:
            isPredatory:
              type: boolean
              description: True if APR > 50%
            severity:
              type: string
              enum: [warning, critical]
            message:
              type: string
    LoanRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        amount:
          type: number
        interestRate:
          type: number
        interestType:
          type: string
          enum: [flat, reducing, compound]
        termLength:
          type: integer
        termUnit:
          type: string
          enum: [weeks, months, years]
        fees:
          type: number
        currency:
          type: string
        results:
          type: object
          properties:
            monthlyPayment:
              type: number
            totalInterest:
              type: number
            totalCost:
              type: number
            effectiveAPR:
              type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SaveLoanRequest:
      allOf:
        - $ref: '#/components/schemas/LoanCalculationRequest'
        - type: object
          properties:
            localId:
              type: string
              description: Client-side ID for offline sync
    LoanHistoryResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/LoanRecord'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ==================
    # SAVINGS SCHEMAS
    # ==================
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input data
            details:
              type: object
              additionalProperties: true
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean